% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_om.R
\name{delta_om}
\alias{delta_om}
\title{Delta om, returns delta omega and updated omega}
\usage{
delta_om(omega, X, e = 0.001, lambda = 0.001)
}
\arguments{
\item{omega}{is a (symmetrical) matrix with the strength of connection between each nodes. Set the the diagonal to zero.}

\item{X}{is a configuration of the network (e.g. if there are three nodes, c(-1,1,-1))}

\item{e}{is the learning parameter (default setting = .001)}

\item{lambda}{is the decay parameter (default setting = .001)}
}
\description{
This function uses Hebb's rule to update omega. It returns the new omega matrix and a matrix with delta omega.
}
