% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Energy.R
\name{energy}
\alias{energy}
\title{Energy}
\usage{
energy(tau, omega, X)
}
\arguments{
\item{tau}{is a vector with a threshold for each node}

\item{omega}{is a (symmetrical) matrix with the strength of connection between each nodes. Set the the diagonal to zero.}

\item{X}{is a configuration of the network (e.g. if there are three nodes, c(-1,1,-1))}
}
\description{
Takes in a configuration of the network, a vector of thresholds, and a matrix with connections.
}
