% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropyS.R
\name{entropyS}
\alias{entropyS}
\title{Gibbs Entropy}
\usage{
entropyS(omega, tau, X, beta = 1, pi_bm)
}
\arguments{
\item{omega}{is a (symmetrical) matrix with the strength of connection between each nodes. Set the the diagonal to zero.}

\item{tau}{is a vector with a threshold for each node}

\item{X}{is the configuration of the network}

\item{beta}{is the dependency parameter}

\item{pi_bm}{is a vector that contains Boltzmann probabilities for all states.}
}
\description{
Calculates Gibbs entropy for a configuration of the network
}
