% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hamiltonian.R
\name{ham}
\alias{ham}
\title{Hamiltonian}
\usage{
ham(tau, omega, X)
}
\arguments{
\item{tau}{is a vector with a threshold for each node}

\item{omega}{is a (symmetrical) matrix with the strength of connection between each nodes. Set the the diagonal to zero.}

\item{X}{is a configuration of the network (e.g. if there are three nodes, c(-1,1,-1))}
}
\description{
Takes in a vector of thresholds, a matrix of interactions, and a configuration of the network
}
