% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIM_simulator.R
\name{LIM_simlator}
\alias{LIM_simlator}
\title{LIM_simulator}
\usage{
LIM_simlator(n = 10, nreps = 100, tau, omega, beta, X1 = sample(c(-1,
  1), n, T), Hebb = T, e = 0.001, lambda = 0.001)
}
\arguments{
\item{n}{is the number of nodes in the network}

\item{nreps}{is the number of iterations}

\item{tau}{is a vector with a threshold for each node}

\item{omega}{is a (symmetrical) matrix with the strength of connection between each nodes. Set the the diagonal to zero.}

\item{beta}{is the dependency parameter in the Ising model}

\item{X1}{is the initial configuration of the network (e.g. if there are three nodes, X1 can be c(-1,1,-1))}

\item{Hebb}{do you want to update omega with Hebb's rule?}

\item{e}{is the learning parameter of Hebb's rule}

\item{lambda}{is the decay parameter of Hebb's rule}
}
\description{
Simulates data from the Learning Ising Model
}
