% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIM_simulator.R
\name{LIM_simulator}
\alias{LIM_simulator}
\title{LIM_simulator}
\usage{
LIM_simulator(n = 10, nreps = 100, tau, omega, beta,
  X1 = sample(c(-1, 1), n, T), Hebb = T, e = 0.001, lambda = 0.001)
}
\arguments{
\item{n}{is the number of nodes in the network}

\item{nreps}{is the number of iterations}

\item{tau}{is a vector with a threshold for each node}

\item{omega}{is a (symmetrical) matrix with the strength of connection between each nodes. Set the the diagonal to zero.}

\item{beta}{is the dependency parameter in the Ising model}

\item{X1}{is the initial configuration of the network (e.g. if there are three nodes, X1 can be c(-1,1,-1))}

\item{Hebb}{do you want to update omega with Hebb's rule?}

\item{e}{is the learning parameter of Hebb's rule}

\item{lambda}{is the decay parameter of Hebb's rule}
}
\value{
configurations contains all the configurations of the network over all nreps iterations

beta contains a vector of length()==nreps with values used for beta

mean omega contains a vector with the mean of omega at each configuration

omega t=... contains omega at 5 timepoints between the first and the last iteration (equal intervals)

gibbs entropy a vector with gibbs entropy for each iteration
}
\description{
Simulates data from the Learning Ising Model
}
