% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIMA_sim1.R
\name{LIMA_sim1}
\alias{LIMA_sim1}
\title{Simulation 1}
\usage{
LIMA_sim1(n = 10, nreps = 5000, tau, omega, beta, X1, Hebb = TRUE)
}
\arguments{
\item{n}{are the number of nodes}

\item{nreps}{are the number of iterations}

\item{tau}{a vector with thresholds}

\item{omega}{a (symmetric) matrix with edge weights (diag==0)}

\item{beta}{is the dependency parameter}

\item{X1}{is the initial configuration of the network}

\item{Hebb}{apply Hebb's rule (logical)}
}
\description{
This function resembles simulation 1 from the LIMA chapter. It generates data for nreps iterations, based on an initial configuration (X1). At each iteration, a random node is flipped with probability p_flip and omega is updated with Hebb's rule (if Hebb=T).
}
