% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flip-probability.R
\name{p_flip}
\alias{p_flip}
\title{Flip-probability}
\usage{
p_flip(tau, omega, X, g, beta = 1)
}
\arguments{
\item{tau}{is a vector with a threshold for each node}

\item{omega}{is a (symmetrical) matrix with the strength of connection between each nodes. Set the the diagonal to zero.}

\item{X}{is a configuration of the network (e.g. if there are three nodes, c(-1,1,-1))}

\item{g}{is a node pciked at random}

\item{beta}{is the dependency parameter in the Ising model}
}
\description{
This function is used to simulate Glauber dynamics calculated the probability a randomly selected node to flip to another state (e.g. -1 -> 1)
}
